sar-audit-engine/requirements.txt
pandas>=2.2.0
pytest>=8.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
weaviate-client>=4.10.0
sentence-transformers>=3.0.0
transformers>=4.46.0
peft>=0.12.0
accelerate>=0.33.0
